Merge unit tests part4
Related-issues: APPLINK-22713, APPLINK-22716

+++ src/components/protocol_handler/test/protocol_handler_tm_test.cc
@@ -364,7 +364,10 @@ TEST_F(ProtocolHandlerImplTest,
  * For ENABLE_SECURITY=OFF session_observer shall be called with protection flag
  * OFF
  */
-TEST_F(ProtocolHandlerImplTest, StartSession_Protected_SessionObserverReject) {
+
+// TODO(OHerasym) : thread qt assert fails
+TEST_F(ProtocolHandlerImplTest,
+       StartSession_Protected_SessionObserverReject) {
   const int call_times = 5;
   AddConnection();
 #ifdef ENABLE_SECURITY
@@ -439,7 +442,10 @@ TEST_F(ProtocolHandlerImplTest,
  * For ENABLE_SECURITY=OFF session_observer shall be called with protection flag
  * OFF
  */
-TEST_F(ProtocolHandlerImplTest, StartSession_Protected_SessionObserverAccept) {
+
+// TODO(OHerasym) : test don't finish executing
+TEST_F(ProtocolHandlerImplTest,
+       StartSession_Protected_SessionObserverAccept) {
   SetProtocolVersion2();
   AddSession();
 }
@@ -448,6 +454,8 @@ TEST_F(ProtocolHandlerImplTest, StartSession_Protected_SessionObserverAccept) {
 /*
  * ProtocolHandler shall send NAck on session_observer rejection
  */
+
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest, EndSession_SessionObserverReject) {
   AddSession();
   const ServiceType service = kRpc;
@@ -472,6 +480,7 @@ TEST_F(ProtocolHandlerImplTest, EndSession_SessionObserverReject) {
 /*
  * ProtocolHandler shall send NAck on wrong hash code
  */
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest, EndSession_Success) {
   AddSession();
   const ServiceType service = kRpc;
@@ -499,7 +508,9 @@ TEST_F(ProtocolHandlerImplTest, EndSession_Success) {
  * ProtocolHandler shall not call Security logics with Protocol version 1
  * Check session_observer with PROTECTION_OFF and Ack with PROTECTION_OFF
  */
-TEST_F(ProtocolHandlerImplTest, SecurityEnable_StartSessionProtocoloV1) {
+// TODO(OHerasym) : test don't finish executing
+TEST_F(ProtocolHandlerImplTest,
+       SecurityEnable_StartSessionProtocoloV1) {
   AddConnection();
   // Add security manager
   AddSecurityManager();
@@ -534,7 +545,9 @@ TEST_F(ProtocolHandlerImplTest, SecurityEnable_StartSessionProtocoloV1) {
  * ProtocolHandler shall not call Security logics on start session with
  * PROTECTION_OFF
  */
-TEST_F(ProtocolHandlerImplTest, SecurityEnable_StartSessionUnprotected) {
+// TODO(OHerasym) : test don't finish executing
+TEST_F(ProtocolHandlerImplTest,
+       SecurityEnable_StartSessionUnprotected) {
   AddConnection();
   // Add security manager
   AddSecurityManager();
@@ -561,7 +574,9 @@ TEST_F(ProtocolHandlerImplTest, SecurityEnable_StartSessionUnprotected) {
 /*
  * ProtocolHandler shall send Ack with PROTECTION_OFF on fail SLL creation
  */
-TEST_F(ProtocolHandlerImplTest, SecurityEnable_StartSessionProtected_Fail) {
+// TODO(OHerasym) : test don't finish executing
+TEST_F(ProtocolHandlerImplTest,
+       SecurityEnable_StartSessionProtected_Fail) {
   AddConnection();
   AddSecurityManager();
   const ServiceType start_service = kRpc;
@@ -594,6 +609,7 @@ TEST_F(ProtocolHandlerImplTest, SecurityEnable_StartSessionProtected_Fail) {
  * ProtocolHandler shall send Ack with PROTECTION_ON on already established and
  * initialized SLLContext
  */
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest,
        SecurityEnable_StartSessionProtected_SSLInitialized) {
   AddConnection();
@@ -637,6 +653,7 @@ TEST_F(ProtocolHandlerImplTest,
 /*
  * ProtocolHandler shall send Ack with PROTECTION_OFF on session handshhake fail
  */
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest,
        SecurityEnable_StartSessionProtected_HandshakeFail) {
   AddConnection();
@@ -703,6 +720,7 @@ TEST_F(ProtocolHandlerImplTest,
  * ProtocolHandler shall send Ack with PROTECTION_ON on session handshhake
  * success
  */
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest,
        SecurityEnable_StartSessionProtected_HandshakeSuccess) {
   AddConnection();
@@ -841,8 +859,10 @@ TEST_F(
  * ProtocolHandler shall send Ack with PROTECTION_ON on session handshhake
  * success
  */
-TEST_F(ProtocolHandlerImplTest,
-       SecurityEnable_StartSessionProtected_HandshakeSuccess_SSLIsNotPending) {
+// TODO(OHerasym) : test don't finish executing
+TEST_F(
+    ProtocolHandlerImplTest,
+    SecurityEnable_StartSessionProtected_HandshakeSuccess_SSLIsNotPending) {
   AddConnection();
   AddSecurityManager();
   const ServiceType start_service = kRpc;
@@ -910,6 +930,7 @@ TEST_F(ProtocolHandlerImplTest,
 }
 #endif  // ENABLE_SECURITY
 
+// TODO (OHerasym) : some times exception on Windows platform
 TEST_F(ProtocolHandlerImplTest, FloodVerification) {
   const size_t period_msec = 10000;
   const size_t max_messages = 1000;
@@ -1009,6 +1030,7 @@ TEST_F(ProtocolHandlerImplTest, FloodVerification_AudioFrameSkip) {
                   &some_data[0]);
   }
 }
+// TODO(OHerasym) : thread qt assert fails
 TEST_F(ProtocolHandlerImplTest, FloodVerificationDisable) {
   const size_t period_msec = 0;
   const size_t max_messages = 0;
@@ -1031,6 +1053,7 @@ TEST_F(ProtocolHandlerImplTest, FloodVerificationDisable) {
   }
 }
 
+// TODO(OHerasym) : thread qt assert fails
 TEST_F(ProtocolHandlerImplTest, MalformedVerificationDisable) {
   const size_t period_msec = 10000;
   const size_t max_messages = 100;
@@ -1096,7 +1119,8 @@ TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification) {
   }
 }
 
-TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_MalformedStock) {
+TEST_F(ProtocolHandlerImplTest,
+       MalformedLimitVerification_MalformedStock) {
   const size_t period_msec = 10000;
   const size_t max_messages = 100;
   InitProtocolHandlerImpl(0u, 0u, true, period_msec, max_messages);
@@ -1160,7 +1184,9 @@ TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_MalformedStock) {
   }
 }
 
-TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_MalformedOnly) {
+// TODO(OHerasym) : thread qt assert fails
+TEST_F(ProtocolHandlerImplTest,
+       MalformedLimitVerification_MalformedOnly) {
   const size_t period_msec = 10000;
   const size_t max_messages = 100;
   InitProtocolHandlerImpl(0u, 0u, true, period_msec, max_messages);
@@ -1214,7 +1240,9 @@ TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_MalformedOnly) {
   }
 }
 
-TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_NullTimePeriod) {
+// TODO(OHerasym) : thread qt assert fails
+TEST_F(ProtocolHandlerImplTest,
+       MalformedLimitVerification_NullTimePeriod) {
   const size_t period_msec = 0;
   const size_t max_messages = 1000;
   InitProtocolHandlerImpl(0u, 0u, true, period_msec, max_messages);
@@ -1240,6 +1268,8 @@ TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_NullTimePeriod) {
                   &some_data[0]);
   }
 }
+
+// TODO(OHerasym) : thread qt assert fails
 TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_NullCount) {
   const size_t period_msec = 10000;
   const size_t max_messages = 0;
@@ -1267,6 +1297,7 @@ TEST_F(ProtocolHandlerImplTest, MalformedLimitVerification_NullCount) {
   }
 }
 
+// TODO(OHerasym) : thread qt assert fails
 TEST_F(ProtocolHandlerImplTest,
        SendEndServicePrivate_NoConnection_MessageNotSent) {
   // Expect check connection with ProtocolVersionUsed
@@ -1279,7 +1310,9 @@ TEST_F(ProtocolHandlerImplTest,
   protocol_handler_impl->SendEndSession(connection_id, session_id);
 }
 
-TEST_F(ProtocolHandlerImplTest, SendEndServicePrivate_EndSession_MessageSent) {
+// TODO(OHerasym) : test don't finish executing
+TEST_F(ProtocolHandlerImplTest,
+       SendEndServicePrivate_EndSession_MessageSent) {
   // Arrange
   AddSession();
   // Expect check connection with ProtocolVersionUsed
@@ -1287,15 +1320,17 @@ TEST_F(ProtocolHandlerImplTest, SendEndServicePrivate_EndSession_MessageSent) {
               ProtocolVersionUsed(connection_id, session_id, _))
       .WillOnce(Return(true));
   // Expect send End Service
-  EXPECT_CALL(
-      transport_manager_mock,
-      SendMessageToDevice(ExpectedMessage(
-          FRAME_TYPE_CONTROL, FRAME_DATA_END_SERVICE, PROTECTION_OFF, kRpc)))
-      .WillOnce(Return(E_SUCCESS));
+  EXPECT_CALL(transport_manager_mock,
+              SendMessageToDevice(
+                  ExpectedMessage(static_cast<int>(FRAME_TYPE_CONTROL),
+                                  static_cast<int>(FRAME_DATA_END_SERVICE),
+                                  PROTECTION_OFF,
+                                  kRpc))).WillOnce(Return(E_SUCCESS));
   // Act
   protocol_handler_impl->SendEndSession(connection_id, session_id);
 }
 
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest,
        SendEndServicePrivate_ServiceTypeControl_MessageSent) {
   // Arrange
@@ -1306,15 +1341,16 @@ TEST_F(ProtocolHandlerImplTest,
       .WillOnce(Return(true));
   // Expect send End Service
   EXPECT_CALL(transport_manager_mock,
-              SendMessageToDevice(ExpectedMessage(FRAME_TYPE_CONTROL,
-                                                  FRAME_DATA_END_SERVICE,
-                                                  PROTECTION_OFF,
-                                                  kControl)))
-      .WillOnce(Return(E_SUCCESS));
+              SendMessageToDevice(ExpectedMessage(
+                  static_cast<int>(FRAME_TYPE_CONTROL),
+                  static_cast<int>(FRAME_DATA_END_SERVICE),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kControl)))).WillOnce(Return(E_SUCCESS));
   // Act
   protocol_handler_impl->SendEndService(connection_id, session_id, kControl);
 }
 
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest, SendHeartBeat_NoConnection_NotSent) {
   // Expect check connection with ProtocolVersionUsed
   EXPECT_CALL(session_observer_mock,
@@ -1326,6 +1362,7 @@ TEST_F(ProtocolHandlerImplTest, SendHeartBeat_NoConnection_NotSent) {
   protocol_handler_impl->SendHeartBeat(connection_id, session_id);
 }
 
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest, SendHeartBeat_Successful) {
   // Arrange
   AddSession();
@@ -1334,15 +1371,17 @@ TEST_F(ProtocolHandlerImplTest, SendHeartBeat_Successful) {
               ProtocolVersionUsed(connection_id, session_id, _))
       .WillOnce(Return(true));
   // Expect send HeartBeat
-  EXPECT_CALL(
-      transport_manager_mock,
-      SendMessageToDevice(ExpectedMessage(
-          FRAME_TYPE_CONTROL, FRAME_DATA_HEART_BEAT, PROTECTION_OFF, kControl)))
-      .WillOnce(Return(E_SUCCESS));
+  EXPECT_CALL(transport_manager_mock,
+              SendMessageToDevice(ExpectedMessage(
+                  static_cast<int>(FRAME_TYPE_CONTROL),
+                  static_cast<int>(FRAME_DATA_HEART_BEAT),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kControl)))).WillOnce(Return(E_SUCCESS));
   // Act
   protocol_handler_impl->SendHeartBeat(connection_id, session_id);
 }
 
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest, SendHeartBeatAck_Successful) {
   // Arrange
   AddSession();
@@ -1353,17 +1392,19 @@ TEST_F(ProtocolHandlerImplTest, SendHeartBeatAck_Successful) {
           DoAll(SetArgReferee<2>(PROTOCOL_VERSION_3), Return(true)));
   // Expect send HeartBeatAck
   EXPECT_CALL(transport_manager_mock,
-              SendMessageToDevice(ExpectedMessage(FRAME_TYPE_CONTROL,
-                                                  FRAME_DATA_HEART_BEAT_ACK,
-                                                  PROTECTION_OFF,
-                                                  kControl)))
-      .WillOnce(Return(E_SUCCESS));
+              SendMessageToDevice(ExpectedMessage(
+                  static_cast<int>(FRAME_TYPE_CONTROL),
+                  static_cast<int>(FRAME_DATA_HEART_BEAT_ACK),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kControl)))).WillOnce(Return(E_SUCCESS));
   // Act
   SendControlMessage(
       PROTECTION_OFF, kControl, session_id, FRAME_DATA_HEART_BEAT);
 }
 
-TEST_F(ProtocolHandlerImplTest, SendHeartBeatAck_WrongProtocolVersion_NotSent) {
+// TODO(OHerasym) : test don't finish executing
+TEST_F(ProtocolHandlerImplTest,
+       SendHeartBeatAck_WrongProtocolVersion_NotSent) {
   // Arrange
   AddSession();
   // Expect two checks of connection and protocol version with
@@ -1374,10 +1415,11 @@ TEST_F(ProtocolHandlerImplTest, SendHeartBeatAck_WrongProtocolVersion_NotSent) {
           DoAll(SetArgReferee<2>(PROTOCOL_VERSION_1), Return(true)));
   // Expect not send HeartBeatAck
   EXPECT_CALL(transport_manager_mock,
-              SendMessageToDevice(ExpectedMessage(FRAME_TYPE_CONTROL,
-                                                  FRAME_DATA_HEART_BEAT_ACK,
-                                                  PROTECTION_OFF,
-                                                  kControl))).Times(0);
+              SendMessageToDevice(
+                  ExpectedMessage(static_cast<int>(FRAME_TYPE_CONTROL),
+                                  static_cast<int>(FRAME_DATA_HEART_BEAT_ACK),
+                                  static_cast<int>(PROTECTION_OFF),
+                                  static_cast<int>(kControl)))).Times(0);
   // Act
   SendControlMessage(
       PROTECTION_OFF, kControl, session_id, FRAME_DATA_HEART_BEAT);
@@ -1385,6 +1427,7 @@ TEST_F(ProtocolHandlerImplTest, SendHeartBeatAck_WrongProtocolVersion_NotSent) {
       PROTECTION_OFF, kControl, session_id, FRAME_DATA_HEART_BEAT);
 }
 
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest,
        SendMessageToMobileApp_SendSingleControlMessage) {
   // Arrange
@@ -1406,15 +1449,17 @@ TEST_F(ProtocolHandlerImplTest,
       .WillOnce(Return(&ssl_context_mock));
 #endif  // ENABLE_SECURITY
   // Expect send message to mobile
-  EXPECT_CALL(
-      transport_manager_mock,
-      SendMessageToDevice(ExpectedMessage(
-          FRAME_TYPE_SINGLE, FRAME_DATA_SINGLE, PROTECTION_OFF, kControl)))
-      .WillOnce(Return(E_SUCCESS));
+  EXPECT_CALL(transport_manager_mock,
+              SendMessageToDevice(ExpectedMessage(
+                  static_cast<int>(FRAME_TYPE_SINGLE),
+                  static_cast<int>(FRAME_DATA_SINGLE),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kControl)))).WillOnce(Return(E_SUCCESS));
   // Act
   protocol_handler_impl->SendMessageToMobileApp(message, is_final);
 }
 
+// TODO(OHerasym) : test don't finish executing
 TEST_F(ProtocolHandlerImplTest,
        SendMessageToMobileApp_SendSingleNonControlMessage) {
   // Arrange
@@ -1440,13 +1485,16 @@ TEST_F(ProtocolHandlerImplTest,
   // Expect send message to mobile
   EXPECT_CALL(transport_manager_mock,
               SendMessageToDevice(ExpectedMessage(
-                  FRAME_TYPE_SINGLE, FRAME_DATA_SINGLE, PROTECTION_OFF, kRpc)))
-      .WillOnce(Return(E_SUCCESS));
+                  static_cast<int>(FRAME_TYPE_SINGLE),
+                  static_cast<int>(FRAME_DATA_SINGLE),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kRpc)))).WillOnce(Return(E_SUCCESS));
   // Act
   protocol_handler_impl->SendMessageToMobileApp(message, is_final);
 }
 
-TEST_F(ProtocolHandlerImplTest, SendMessageToMobileApp_SendMultiframeMessage) {
+TEST_F(ProtocolHandlerImplTest,
+       SendMessageToMobileApp_SendMultiframeMessage) {
   // Arrange
   AddSession();
   const bool is_final = true;
@@ -1471,20 +1519,22 @@ TEST_F(ProtocolHandlerImplTest, SendMessageToMobileApp_SendMultiframeMessage) {
   // Expect sending message frame by frame to mobile
   EXPECT_CALL(transport_manager_mock,
               SendMessageToDevice(ExpectedMessage(
-                  FRAME_TYPE_FIRST, FRAME_DATA_FIRST, PROTECTION_OFF, kBulk)))
-      .WillOnce(Return(E_SUCCESS));
+                  static_cast<int>(FRAME_TYPE_FIRST),
+                  static_cast<int>(FRAME_DATA_FIRST),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kBulk)))).WillOnce(Return(E_SUCCESS));
   EXPECT_CALL(transport_manager_mock,
-              SendMessageToDevice(ExpectedMessage(FRAME_TYPE_CONSECUTIVE,
-                                                  first_consecutive_frame,
-                                                  PROTECTION_OFF,
-                                                  kBulk)))
-      .WillOnce(Return(E_SUCCESS));
+              SendMessageToDevice(ExpectedMessage(
+                  static_cast<int>(FRAME_TYPE_CONSECUTIVE),
+                  static_cast<int>(first_consecutive_frame),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kBulk)))).WillOnce(Return(E_SUCCESS));
   EXPECT_CALL(transport_manager_mock,
-              SendMessageToDevice(ExpectedMessage(FRAME_TYPE_CONSECUTIVE,
-                                                  FRAME_DATA_LAST_CONSECUTIVE,
-                                                  PROTECTION_OFF,
-                                                  kBulk)))
-      .WillOnce(Return(E_SUCCESS));
+              SendMessageToDevice(ExpectedMessage(
+                  static_cast<int>(FRAME_TYPE_CONSECUTIVE),
+                  static_cast<int>(FRAME_DATA_LAST_CONSECUTIVE),
+                  static_cast<int>(PROTECTION_OFF),
+                  static_cast<int>(kBulk)))).WillOnce(Return(E_SUCCESS));
   // Act
   protocol_handler_impl->SendMessageToMobileApp(message, is_final);
 }
